docs/tools/requirements.txt
GitPython==3.1.40
natsort==8.4.0
